requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Header-Session==0.0.7
Flask-HTTPAuth==4.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
PyJWT==2.8.0
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
